Unet_Docker/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.4.2
